+++ NEWS.md
@@ -1,4 +1,5 @@
 ##### Version 0.7.2
+* Handle sparse data of class Matrix::dgCMatrix
 * Add prediction of standard errors to predict()
 
 ##### Version 0.7.1